+++ src/main/org/firebirdsql/jdbc/AbstractCallableStatement.java
@@ -306,8 +306,8 @@ public abstract class AbstractCallableStatement extends FBPreparedStatement impl
         final boolean hasResultSet = super.internalExecute(sendOutParams);
         if (hasResultSet && isSingletonResult) {
             // Safeguarding first row so it will work even if the result set from getResultSet is manipulated
-            singletonRs = new FBResultSet(fbStatement.getRowDescriptor(), connection,
-                    new ArrayList<>(specialResult), true);
+            singletonRs = new FBResultSet(fbStatement.getRowDescriptor(), connection, new ArrayList<>(specialResult),
+                    null, true, false);
         }
         return hasResultSet;
     }