+++ src/test/org/firebirdsql/jdbc/FBPreparedStatementTest.java
@@ -1449,6 +1449,25 @@ public class FBPreparedStatementTest extends FBJUnit4TestBase {
         }
     }
 
+    /**
+     * Tests for <a href="https://github.com/FirebirdSQL/jaybird/issues/729">jaybird#729</a>.
+     */
+    @Test
+    public void preparedStatementExecuteProcedureShouldNotTrim_729() throws Exception {
+        executeDDL(con,
+                "create procedure char_return returns (val char(5)) as\n" +
+                "begin\n" +
+                "  val = 'A';\n" +
+                "end");
+
+        try (PreparedStatement stmt = con.prepareStatement("execute procedure char_return");
+             ResultSet rs = stmt.executeQuery()) {
+            assertTrue("Expected a row", rs.next());
+            assertEquals("Unexpected trim by getObject", "A    ", rs.getObject(1));
+            assertEquals("Unexpected trim by getString", "A    ", rs.getString(1));
+        }
+    }
+
     private void prepareTestData() throws SQLException {
         con.setAutoCommit(false);
         try (PreparedStatement pstmt = con.prepareStatement(INSERT_DATA)) {